Cthonios

Documentation for Cthonios.

Cthonios.CthoniosCommon

Cthonios.CthoniosCommon

Cthonios.CthoniosLogger

Cthonios.CthoniosLogger

Cthonios.CthoniosTimer

Cthonios.CthoniosTimer

Cthonios.PostProcessor

Cthonios.end_section

Cthonios.load_input_file

Cthonios.minimize_trust_region_sub_problem!

Cthonios.new_section

Cthonios.parse_boundary_conditions

Cthonios.parse_domains

Cthonios.parse_domains

Cthonios.parse_functions

Cthonios.parse_input_file

Cthonios.parse_linear_solvers

Cthonios.parse_materials

Cthonios.parse_mesh

Cthonios.parse_nonlinear_solvers

Cthonios.parse_problems

Cthonios.parse_sections

Cthonios.parse_time_stepper

Cthonios.update_unknown_dofs!

Cthonios.with_logger

Cthonios.with_logger

Cthonios.CthoniosCommon — Type

logger::Cthonios.CthoniosLogger

timer::Cthonios.CthoniosTimer

backend::Cthonios.CthoniosBackend

Cthonios.CthoniosCommon — Method

CthoniosCommon(
    file_name::String,
    backend::Cthonios.CthoniosBackend
) -> Cthonios.CthoniosCommon

CthoniosCommon(file_name, backend)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:86.

Cthonios.CthoniosLogger — Type

logger::LoggingExtras.FormatLogger

Cthonios.CthoniosLogger — Method

CthoniosLogger(file_name::String) -> Cthonios.CthoniosLogger

CthoniosLogger(file_name)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:30.

Cthonios.CthoniosTimer — Type

timer::TimerOutputs.TimerOutput

Cthonios.CthoniosTimer — Method

CthoniosTimer() -> Cthonios.CthoniosTimer

CthoniosTimer()

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:63.

Cthonios.PostProcessor — Method

Constructor for post processor

closes exodus db by default, needs to be re-opend downstream this is mainly for debuggin purposes

Cthonios.end_section — Method

end_section(section_name::String)

end_section(section_name)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:126.

Cthonios.load_input_file — Method

load_input_file(input_file::String) -> Dict{Symbol, Any}

This method is a barrier method for type stability

load_input_file(input_file)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:4.

Cthonios.minimize_trust_region_sub_problem! — Method

minimize_trust_region_sub_problem!(
    z,
    solver::Cthonios.TrustRegionSolver,
    domain,
    common::Cthonios.CthoniosCommon,
    x::AbstractVector,
    r_in,
    tr_size::Float64
) -> Tuple{Symbol, Int64}

minimize r * z + 0.5 * z * J * z

minimize_trust_region_sub_problem!(
    z,
    solver,
    domain,
    common,
    x,
    r_in,
    tr_size
)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/solvers/TrustRegionSolver.jl:187.

Cthonios.new_section — Method

new_section(section_name::String)

new_section(section_name)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:114.

Cthonios.parse_boundary_conditions — Method

parse_boundary_conditions(
    input_settings::Dict{Symbol, Any},
    domain_key::Symbol
) -> Dict{Symbol, Any}

Barrier for type stability. Returns inputsettings[:boundaryconditions] but also replaces function names with function input settings.

parse_boundary_conditions(input_settings, domain_key)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:12.

Cthonios.parse_domains — Method

parse_domains(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Helper to parse a domain and all settings need to set it up

parse_domains(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:98.

Cthonios.parse_domains — Method

parse_domains(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Helper to parse one domain

parse_domains(input_settings, key)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:86.

parse_domains(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:98.

Cthonios.parse_functions — Method

parse_functions(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Barrier for type stability. Simply returns input_settings[:functions]

parse_functions(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:28.

Cthonios.parse_input_file — Method

parse_input_file(input_file::String) -> Dict{Symbol, Any}

Main entry point for parsing

parse_input_file(input_file)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:133.

Cthonios.parse_linear_solvers — Method

parse_linear_solvers(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Barrier for type stability. Simply returns input_settings[Symbol("linear solvers")]

parse_linear_solvers(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:35.

Cthonios.parse_materials — Method

parse_materials(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Barrier for type stability. Simply returns input_settings[:materials]

parse_materials(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:42.

Cthonios.parse_mesh — Method

parse_mesh(
    input_settings::Dict{Symbol, Any},
    domain_key::Symbol
) -> Dict{Symbol, String}

Barrier for type stability. Simply returns input_settings[:mesh]

parse_mesh(input_settings, domain_key)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:56.

Cthonios.parse_nonlinear_solvers — Method

parse_nonlinear_solvers(
    input_settings::Dict{Symbol, Any}
) -> Dict{Symbol, Any}

Barrier for type stability. Simply returns input_settings[Symbol("nonlinear solvers")]

parse_nonlinear_solvers(input_settings)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:49.

Cthonios.parse_problems — Method

parse_problems(
    input_settings::Dict{Symbol, Any},
    domain_settings::Dict{Symbol, Any},
    linear_solver_settings::Dict{Symbol, Any},
    nonlinear_solver_settings::Dict{Symbol, Any}
) -> Any

Helper to parse problems

parse_problems(
    input_settings,
    domain_settings,
    linear_solver_settings,
    nonlinear_solver_settings
)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:109.

Cthonios.parse_sections — Method

parse_sections(
    input_settings::Dict{Symbol, Any},
    domain_key::Symbol
) -> Vector{Dict{Symbol, Any}}

Barrier for type stability. Returns input_settings[:sections] but replaces material names with material definitions

parse_sections(input_settings, domain_key)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:64.

Cthonios.parse_time_stepper — Method

parse_time_stepper(
    input_settings::Dict{Symbol, Any},
    domain_key::Symbol
) -> Dict{Symbol, Any}

Barrier for type stability. Simple returns input_settings[Symbol("time stepper")]

parse_time_stepper(input_settings, domain_key)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Parsers.jl:76.

Cthonios.update_unknown_dofs! — Method

update_unknown_dofs!(d::Cthonios.QuasiStaticDomain) -> Any

This methods assumes the bc dofs and func ids are already properly set in the domain coming in

update_unknown_dofs!(d)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/domains/QuasiStaticDomain.jl:120.

Cthonios.with_logger — Method

with_logger(f, common::Cthonios.CthoniosCommon) -> Any

with_logger(f, common)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:96.

Cthonios.with_logger — Method

with_logger(f, logger::Cthonios.CthoniosLogger) -> Any

with_logger(f, logger)

defined at /home/runner/work/Cthonios.jl/Cthonios.jl/src/Common.jl:45.