Domains

A domain in Cthonios is a group of element sections and all the associated datastructures necessary to run a finite element calculation. Different types of domains are used for different problems, e.g. quasi-static vs. dynamics problems have different requirements and solution strategies.

Input file syntax example

domains:
  domain_1:
    mesh:
      ...
    boundary conditions:
      ...
    sections:
      ...
    time stepper:
      ...

Quasi-static Domain

Quasi-static domains are appropriate for problems where the following equation holds

$\nabla\cdot \mathbf{P} + \mathbf{b} = \mathbf{0},$ where $\mathbf{P}$ is the first Piola-Kirchoff stress, $\nabla$ is the gradient operator with respsect to the reference corrdinates, and $\mathbf{b}$ is a body force.

Index

Cthonios.QuasiStaticDomain — Type

Cthonios.QuasiStaticDomain — Method

QuasiStaticDomain(
    input_settings::Dict{Symbol, Any}
) -> Cthonios.QuasiStaticDomain{Dof, Funcs, Vector{Int64}, Vector{Int64}, Vector{Int64}, Sections, DomainCache} where {Dof<:(FiniteElementContainers.DofManager{Int64, _A, _B, Vector{Float64}, _C} where {_A, _B, _C<:AbstractVector{Int64}}), Funcs<:(Vector), Sections<:NamedTuple, DomainCache<:(Cthonios.QuasiStaticDomainCache{Time, V1, _A, V3, _B, _C, Vector{Float64}} where {Time<:Cthonios.ConstantTimeStepper, V1<:(FiniteElementContainers.VectorizedNodalField{_A, 2, _B, _C, _D} where {_A, _B, _C, _D<:AbstractVector{_A}}), _A, V3<:(FiniteElementContainers.VectorizedNodalField{Float64, 2, _A, _B, Vector{Float64}} where {_A, _B}), _B, _C})}

Cthonios.QuasiStaticDomain — Method

QuasiStaticDomain(
    input_file::String,
    key
) -> Cthonios.QuasiStaticDomain{Dof, Funcs, Vector{Int64}, Vector{Int64}, Vector{Int64}, Sections, DomainCache} where {Dof<:(FiniteElementContainers.DofManager{Int64, _A, _B, Vector{Float64}, _C} where {_A, _B, _C<:AbstractVector{Int64}}), Funcs<:(Vector), Sections<:NamedTuple, DomainCache<:(Cthonios.QuasiStaticDomainCache{Time, V1, _A, V3, _B, _C, Vector{Float64}} where {Time<:Cthonios.ConstantTimeStepper, V1<:(FiniteElementContainers.VectorizedNodalField{_A, 2, _B, _C, _D} where {_A, _B, _C, _D<:AbstractVector{_A}}), _A, V3<:(FiniteElementContainers.VectorizedNodalField{Float64, 2, _A, _B, Vector{Float64}} where {_A, _B}), _B, _C})}

Cthonios.QuasiStaticDomainCache — Type

time::Any

X::Any

Uu::Any

ΔUu::Any

U::Any

props::Any

state_old::Any

state_new::Any

Π::Any

Πs::Any

f::Any

V::Any

Hv::Any

Base.similar — Method

similar(
    cache::Cthonios.QuasiStaticDomainCache
) -> Cthonios.QuasiStaticDomainCache

Cthonios.step! — Method

step!(domain::Cthonios.QuasiStaticDomain) -> Int64

Advances the domain's time stepper one step.

Cthonios.unpack — Method

unpack(
    cache::Cthonios.QuasiStaticDomainCache
) -> NTuple{13, Any}

Cthonios.update_bcs! — Method

update_bcs!(U, domain::Cthonios.QuasiStaticDomain, Xs)

This method updates the nodes with dirichlet bcs in U

Cthonios.update_fields! — Method

update_fields!(
    U,
    domain::Cthonios.QuasiStaticDomain,
    Xs,
    Uu
) -> AbstractVector{<:Number}

Updates both the bcs and unknowns in U based on the current free vs. fixed dof layout in domain.dof and the bcs currently in the domain.

I think this one is deprecated?

Cthonios.update_unknown_dofs! — Method

update_unknown_dofs!(d::Cthonios.QuasiStaticDomain) -> Any

This methods assumes the bc dofs and func ids are already properly set in the domain coming in

Cthonios.update_unknowns! — Method

update_unknowns!(
    U,
    domain::Cthonios.QuasiStaticDomain,
    Uu
) -> AbstractVector{<:Number}

Updates the unknowns in U with the values in Uu based on the current free vs. fixed dof layout in domain.dof.

I think this one is deprecated?