Functions

Functions can be used to help drive boundary conditions, initial conditions, etc.

Index

Types

Cthonios.ScalarFunction — Type

mutable struct FunctionWrapper{T1, Tuple{StaticArraysCore.SArray{Tuple{D}, T2, 1, D}, T3}}

Cthonios.ScalarFunctionWrapperArray — Type

Array{T,N} <: AbstractArray{T,N}

N-dimensional dense array with elements of type T.

Useful methods

Cthonios.setup_function — Function

setup_function(inputs::Dict{Symbol, Any}) -> Any

This method constructs a Function from a a dictionary of inputs. Typically this will be called from setup_functions which will recieve a vector of dictionaries from a parsed yaml file.

The syntax in the input file should mimic the following

bc_name:
  type: ScalarFunction{NDIM, Float64, Float64, Float64}
  expression: (x, t) -> 1.0 * t

Cthonios.setup_functions — Function

setup_functions(inputs::Vector{Dict{Symbol, Any}}) -> Vector

Input file syntax

functions:
  bc_name_1:
    type: ScalarFunction{NDIM, Float64, Float64, Float64}
    expression: (x, t) -> 1.0 * t
  bc_name_2:
    type: ScalarFunction{NDIM, Float64, Float64, Float64}
    expression: (x, t) -> 1.0 * x * t
  ...
  bc_name_n:
    type: ScalarFunction{NDIM, Float64, Float64, Float64}
    expression: (x, t) -> 0.0