Iterators

Cthonios.domain_iterator! — Method

domain_iterator!(
    global_val,
    f,
    domain::Domain,
    Uu,
    p::ObjectiveParameters
)

Iterator over a domain domain to fill a global value global_val based on a quadrature level function f provided a nodal field U and set of paramaters p. This method is useful for filling quantities such as objectives, gradients, or hessians.

Cthonios.element_coordinates — Method

element_coordinates(section, X, e) -> Any

Cthonios.element_fields — Method

element_fields(section, U, e) -> Any

Cthonios.scratch_variable — Method

scratch_variable(
    global_val::FiniteElementContainers.Assembler,
    section
) -> Any

Setup a scratch variable for a stiffness like calculation

Cthonios.scratch_variable — Method

scratch_variable(global_val::Vector, section) -> Any

Setup a scratch variable for an energy like calculation

Cthonios.scratch_variable — Method

scratch_variable(
    global_val::Union{FiniteElementContainers.NodalField, Matrix},
    section
) -> Any

Setup a scratch variable for a force like calculation