Nonlinear Solvers

One of our goals is to make adding new and hacking existing nonlinear solvers easy!

Input File Syntax Example

Below is an example input file block for a trust region solver that leverages a direct solver with an LDLT factorization. The nonlinear solver is also utilizing warm start to improve the first few iterations of each load step.

linear solvers:
  direct:
    type: DirectLinearSolver
    factorization method: ldl

nonlinear solvers:
  trs:
    type: TrustRegionSolver
    linear solver: direct
    warm start: on

Newton Solver

Cthonios.NewtonSolver — Type

settings::Cthonios.NewtonSolverSettings

linear_solver::Cthonios.AbstractLinearSolver

use_warm_start::Bool

Cthonios.NewtonSolver — Method

NewtonSolver(
    input_settings::Dict,
    domain::Cthonios.QuasiStaticDomain,
    backend
) -> Cthonios.NewtonSolver{Cthonios.NewtonSolverSettings}

Cthonios.NewtonSolverSettings — Type

max_steps::Int64

relative_tolerance::Float64

absolute_tolerance::Float64

Cthonios.NewtonSolverSettings — Method

NewtonSolverSettings(
    input_settings::Dict
) -> Cthonios.NewtonSolverSettings

Cthonios.solve! — Method

solve!(
    solver::Cthonios.NewtonSolver,
    domain::Cthonios.QuasiStaticDomain,
    common::Cthonios.CthoniosCommon
)

Trust Region Solver

Cthonios.TrustRegionSolver — Type

settings::Cthonios.TrustRegionSolverSettings

linear_solver::Cthonios.AbstractLinearSolver

cauchy_point::AbstractVector

q_newton_point::AbstractVector

d::AbstractVector

y_scratch_1::AbstractVector

y_scratch_2::AbstractVector

y_scratch_3::AbstractVector

y_scratch_4::AbstractVector

use_warm_start::Bool

Cthonios.TrustRegionSolver — Method

TrustRegionSolver(
    input_settings::Dict{Symbol, Any},
    domain::Cthonios.QuasiStaticDomain,
    backend
) -> Cthonios.TrustRegionSolver{Cthonios.TrustRegionSolverSettings}

Cthonios.TrustRegionSolverSettings — Type

t_1::Float64

t_2::Float64

η_1::Float64

η_2::Float64

η_3::Float64

max_trust_iters::Int64

tol::Float64

max_cg_iters::Int64

max_cumulative_cg_iters::Int64

cg_tol::Float64

cg_inexact_solve_ratio::Float64

tr_size::Float64

min_tr_size::Float64

check_stability::Bool

use_preconditioned_inner_product_for_cg::Bool

use_incremental_objective::Bool

debug_info::Bool

over_iter::Int64

Cthonios.calculate_cauchy_point! — Method

calculate_cauchy_point!(
    cauchy_point,
    solver,
    common,
    g,
    Hg,
    tr_size
) -> Any

Cthonios.dog_leg_step! — Method

dog_leg_step!(
    d,
    solver,
    common,
    cauchy_point,
    q_newton_point,
    tr_size
)

Cthonios.minimize_trust_region_sub_problem! — Method

minimize_trust_region_sub_problem!(
    z,
    solver::Cthonios.TrustRegionSolver,
    domain,
    common::Cthonios.CthoniosCommon,
    x::AbstractVector,
    r_in,
    tr_size::Float64
) -> Tuple{Symbol, Int64}

minimize r * z + 0.5 * z * J * z

Cthonios.preconditioner! — Method

preconditioner!(solver, domain, common, Uu)

Cthonios.solve! — Method

solve!(
    solver::Cthonios.TrustRegionSolver,
    domain::Cthonios.QuasiStaticDomain,
    common::Cthonios.CthoniosCommon
)

Cthonios.update_tr_size — Method

update_tr_size(
    solver,
    model_objective,
    real_objective,
    step_type,
    tr_size,
    real_res_norm,
    g_norm
) -> Tuple{Any, Any}

General

Cthonios.NonlinearSolver — Type

Cthonios.NonlinearSolverSettings — Type

Cthonios.gradient — Method

gradient(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u
) -> Any

Cthonios.hessian — Method

hessian(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u
) -> Any

Cthonios.hvp — Method

hvp(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u,
    v
) -> Any

Cthonios.objective — Method

objective(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u
) -> Any

Cthonios.objective_and_gradient — Method

objective_and_gradient(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u
) -> Tuple{Any, Any}

Cthonios.objective_gradient_and_hessian — Method

objective_gradient_and_hessian(
    solver,
    domain::Cthonios.QuasiStaticDomain,
    common,
    u
) -> Tuple{Any, Any, Any}

Cthonios.update_unknown_dofs! — Method

update_unknown_dofs!(
    solver::Cthonios.NonlinearSolver,
    d::Cthonios.QuasiStaticDomain
)