Sections

Within Cthonios, sections refer to groups or "blocks" of elements that share a common element topology and material model. We allow by default for each element to possess different fixed material properties and an arbitrary number of state variables. Since the material model is the same for all elements in a section, the number of state variables will be the same at each element/quadrature point.

Input file syntax

Cthonios.setup_sections — Function

setup_sections(
    input_settings::Vector{Dict{Symbol, Any}},
    mesh::FiniteElementContainers.FileMesh,
    dof
) -> Tuple{NamedTuple, Vararg{Any, 4}}

TotalLagrangeSection

Cthonios.TotalLagrangeSection — Type

block_id::Any

fspace::Any

formulation::Any

model::Any

Section for total Lagrangia (the actual way to spell his name) formulations