Time steppers

Cthonios.AbstractTimeStepper — Type

abstract type AbstractTimeStepper{T}

Abstract type for time steppers. Expects a type T to correspond to the type for time values such as Float64 or a time with a unit from for example Unitful.

Cthonios.ConstantTimeStepper — Type

struct ConstantTimeStepper{T, S<:AbstractArray{T, 1}, I<:(AbstractVector{<:Integer})} <: Cthonios.AbstractTimeStepper{T}

start_time::Any

end_time::Any

current_time::AbstractVector

current_time_step::AbstractVector{<:Integer}

Δt::AbstractVector

Time stepper type with a constant time step. start_time - the initial time value. end_time - time to end the simulation. current_time - the current time stored by this time stepper. current_time_step - the current index of the time step. Δt - the time step to use for all time steps.

Cthonios.ConstantTimeStepper — Method

ConstantTimeStepper(
    start_time::Number,
    end_time::Number,
    Δt::Number
) -> ConstantTimeStepper

Method to construct a ConstantTimeStepper. start_time - the initial time value. end_time - time to end the simulation. Δt - the time step to use for all time steps.

Cthonios.step! — Method

step!(time::ConstantTimeStepper)

Method to increment time.current_time by Δt.