Domains

A domain in Cthonios is a group of element sections and all the associated datastructures necessary to run a finite element calculation. Different types of domains are used for different problems, e.g. quasi-static vs. dynamics problems have different requirements and solution strategies.

Input file syntax example

domains:
  domain_1:
    mesh:
      ...
    boundary conditions:
      ...
    sections:
      ...
    time stepper:
      ...

Quasi-static Domain

Quasi-static domains are appropriate for problems where the following equation holds

$\nabla\cdot \mathbf{P} + \mathbf{b} = \mathbf{0},$ where $\mathbf{P}$ is the first Piola-Kirchoff stress, $\nabla$ is the gradient operator with respsect to the reference corrdinates, and $\mathbf{b}$ is a body force.

Index

Cthonios.AbstractDomain — Type

abstract type AbstractDomain

Cthonios.Domain — Type

struct Domain{M, D, S, DBCs, DDofs} <: Cthonios.AbstractDomain

mesh::Any

dof::Any

sections::Any

dirichlet_bcs::Any

dirichlet_dofs::Any

Cthonios.Domain — Method

Domain(
    mesh_file::String,
    sections_in,
    dbcs_in,
    n_dofs::Int64
) -> Domain{M, D, S, _A, Vector{Int64}} where {M<:(FiniteElementContainers.FileMesh{Exodus.ExodusDatabase{M, I, B, F, Init}} where {M, I, B, F, Init}), D<:(FiniteElementContainers.DofManager{Int64, _A, _B, Vector{Float64}, _C} where {_A, _B, _C<:(AbstractArray{Int64})}), S<:NamedTuple, _A}

Cthonios.StaticAssembler — Method

StaticAssembler(
    domain::Domain
) -> FiniteElementContainers.StaticAssembler{Float64, Int64, _A, _B, _C, _D, _E, _F, _G, _H, _I, _J, _K, Vector{Int64}, Vector{Int64}, Vector{Float64}} where {_A<:AbstractVector{Int64}, _B<:AbstractVector{Int64}, _C<:AbstractVector{Int64}, _D<:AbstractVector{Int64}, _E<:AbstractVector{Int64}, _F<:FiniteElementContainers.NodalField, _G<:AbstractVector{Float64}, _H, _I, _J, _K}

some FEMContainers abuse

Cthonios.create_fields — Method

create_fields(domain::Domain) -> Any

Cthonios.create_unknowns — Method

create_unknowns(domain::Domain) -> Any

Cthonios.dirichlet_dofs — Method

dirichlet_dofs(domain::Domain) -> Any

Cthonios.update_field_bcs! — Method

update_field_bcs!(U, domain::Domain, Ubc)

Cthonios.update_unknown_dofs! — Method

update_unknown_dofs!(domain::Domain, asm)

Cthonios.update_unknown_dofs! — Method

update_unknown_dofs!(domain::Domain)