Nonlinear Solvers

One of our goals is to make adding new and hacking existing nonlinear solvers easy!

Input File Syntax Example

Below is an example input file block for a trust region solver that leverages a direct solver with an LDLT factorization. The nonlinear solver is also utilizing warm start to improve the first few iterations of each load step.

linear solvers:
  direct:
    type: DirectLinearSolver
    factorization method: ldl

nonlinear solvers:
  trs:
    type: TrustRegionSolver
    linear solver: direct
    warm start: on

General methods and abstract types

Cthonios.AbstractNonlinearSolver — Type

abstract type AbstractNonlinearSolver{L, O, U, T}

a nonlinear solver needs to define the following methods

check_convergence - returns a bool

logger - @info information

step! - step for the solver

it needs the following types

a linear solver

an objective

an unknown vector

an int called max_iter

Cthonios.create_unknowns — Method

create_unknowns(
    solver::Cthonios.AbstractNonlinearSolver
) -> Any

Creates a set of unknowns for the nonlinear solver

Cthonios.solve! — Method

solve!(solver::Cthonios.AbstractNonlinearSolver, Uu, p)

Generic method to fall back on if step! is defined

Newton Solver

Cthonios.NewtonSolver — Type

struct NewtonSolver{L, O, U, T} <: Cthonios.AbstractNonlinearSolver{L, O, U, T}

linear_solver::Any

objective::Any

ΔUu::Any

timer::Any

max_iter::Int64

abs_tol::Float64

rel_tol::Float64

use_warm_start::Bool

Cthonios.NewtonSolver — Method

NewtonSolver(
    objective::Objective,
    p,
    linear_solver_type,
    timer
) -> NewtonSolver{_A, O} where {_A, O<:Objective}

Cthonios.check_convergence — Method

check_convergence(solver::NewtonSolver, R0_norm) -> Bool

Cthonios.logger — Method

logger(solver::NewtonSolver, n::Int64, norm_R0)

Cthonios.step! — Method

step!(solver::NewtonSolver, Uu, p)