Sections

Within Cthonios, sections refer to groups or "blocks" of elements that share a common element topology and material model. We allow by default for each element to possess different fixed material properties and an arbitrary number of state variables. Since the material model is the same for all elements in a section, the number of state variables will be the same at each element/quadrature point.

Input file syntax

TODO

Cthonios.AbstractSection — Type

abstract type AbstractSection

Abstract base type for sections (both for input and internal).

Cthonios.AbstractSectionInput — Type

abstract type AbstractSectionInput <: Cthonios.AbstractSection

Abstract type for section inputs.

Cthonios.AbstractSectionInternal — Type

abstract type AbstractSectionInternal{F, P} <: Cthonios.AbstractSection

Abstract type for section internals. P corresponds to the physics F corresponds to the formulation

Cthonios.NeumannBCSectionInternal — Type

struct NeumannBCSectionInternal{F, P<:Cthonios.AbstractPhysics, M, B} <: Cthonios.AbstractSectionInternal{F, P<:Cthonios.AbstractPhysics}

block_id::Int64

q_order::Int64

fspace::Any

physics::Cthonios.AbstractPhysics

props::Any

bc::Any

Section internals. physics - Physics object to use in this section. block_name - Name of exodus block to construct section from. fspace - Function space for this element type

Cthonios.Section — Type

struct Section{P, M} <: Cthonios.AbstractSectionInput

block_name::String

q_order::Int64

physics::Any

props::Any

Section inputs. physics - Physics object to use in this section. block_name - Name of exodus block to construct section from. q_order - Quadrature Order to use

Cthonios.SectionInternal — Type

struct SectionInternal{F, P<:Cthonios.AbstractPhysics, M} <: Cthonios.AbstractSectionInternal{F, P<:Cthonios.AbstractPhysics}

block_id::Int64

q_order::Int64

fspace::Any

physics::Cthonios.AbstractPhysics

props::Any

Section internals. physics - Physics object to use in this section. block_name - Name of exodus block to construct section from. fspace - Function space for this element type

Cthonios.SectionInternal — Method

SectionInternal(
    mesh,
    dof::FiniteElementContainers.DofManager,
    section
) -> Cthonios.SectionInternal{F} where F<:(FiniteElementContainers.NonAllocatedFunctionSpace{_A, _B, FiniteElementContainers.VectorizedElementField{T, N, NN, NE, Vals}, _C, ReferenceFiniteElements.ReferenceFE{Itype, Ftype, Etype, SEType, Backend, EdgeNodes, FaceNodes, InteriorNodes, Coords, CellInterps, SurfaceCoords, SurfaceInterps}} where {_A, _B, T, N, NN, NE, Vals<:AbstractVector{T}, _C<:(FiniteElementContainers.ElementField), Itype, Ftype, Etype, SEType, Backend, EdgeNodes, FaceNodes, InteriorNodes, Coords, CellInterps, SurfaceCoords, SurfaceInterps})

Constructor for SectionInternal. mesh - FileMesh object. dof - DofManager object. section - SectionInput object.

Base.size — Method

size(
    section::Cthonios.AbstractSectionInternal
) -> NTuple{4, Any}

Retrieves the sizes of a section in the following tuple (ND, NN, NP, NS) where ND - Number of dimensions NN - Number of nodes per element NP - Number of properties NS - Number of states

Cthonios.num_fields — Method

num_fields(section::Cthonios.AbstractSection) -> Any

Retrieves the number of fields for the physics in section

Cthonios.num_properties — Method

num_properties(section::Cthonios.AbstractSection) -> Any

Retrieves the number of properties for the physics in section

Cthonios.num_states — Method

num_states(section::Cthonios.AbstractSection) -> Any

Retrieves the number of states for the physics in section